Basic Movement Down
This movement makes use of the addforce method to have increasing and decreasing acceleration. This will also make it easy to have the player affected by the objects around them. Still needs to change based on the camera direction though and if we use slopes will need to be changed to make sure we don't slide too much as there is currently nothing stopping our vertical movement from going too high.